dev./prod. versions of store
