+++ dispatch/static/manager/package.json
@@ -46,6 +46,7 @@
     "react-router": "^2.7.0",
     "react-router-redux": "^4.0.5",
     "redux": "^3.5.2",
+    "redux-logger": "^3.0.6",
     "redux-promise-middleware": "^4.0.0",
     "redux-thunk": "^2.1.0",
     "sass-loader": "^6.0.2",
