+++ dispatch/static/manager/src/js/index.js
@@ -1,45 +1,14 @@
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
-import { createStore, combineReducers, applyMiddleware } from 'redux'
-import thunk from 'redux-thunk'
-import { Router, Route, IndexRoute, useRouterHistory } from 'react-router'
-import { syncHistoryWithStore, routerReducer, routerMiddleware } from 'react-router-redux'
-import { createHistory } from 'history'
-import { loadingBarReducer, loadingBarMiddleware } from 'react-redux-loading-bar'
-import promiseMiddleware from 'redux-promise-middleware'
+import { Router, Route, IndexRoute } from 'react-router'
 
 // Base CSS styles
 require('../styles/admin.scss')
 
+import { store, history } from './store'
 import * as Pages from './pages'
 import * as Containers from './containers'
-import appReducer from './reducers/AppReducer'
-import modalReducer from './reducers/ModalReducer'
-
-const BASE_PATH = '/admin'
-
-const browserHistory = useRouterHistory(createHistory)({ basename: BASE_PATH })
-const router = routerMiddleware(browserHistory)
-
-const store = createStore(
-  combineReducers({
-    app: appReducer,
-    routing: routerReducer,
-    loadingBar: loadingBarReducer,
-    modal: modalReducer
-  }),
-  applyMiddleware(
-    thunk,
-    router,
-    promiseMiddleware(),
-    loadingBarMiddleware({
-      promiseTypeSuffixes: ['PENDING', 'FULFILLED', 'REJECTED'],
-    })
-  )
-)
-
-const history = syncHistoryWithStore(browserHistory, store)
 
 render((
   <Provider store={store}>
