+++ dispatch/static/manager/src/js/store/index.js
@@ -1,0 +1,5 @@
+if (process.env.NODE_ENV == 'production') {
+  module.exports = require('./production.js')
+} else {
+  module.exports = require('./development.js')
+}
